Fix builds with pkg-config
The $HOME directory differs between the musescore_deps repository and
the MuseScore repository. The former uses
/Users/runner/work/musescore_deps/musescore_deps/ while the latter
uses /Users/runner/. This leads to a mismatch between the directories in
listed in a .pc pkg-config file created by this repository and the patch
where libraries from the musescore_deps tarball are actually placed
during a MuseScore MacOS CI build.

This does not seem to be an issue for the current libraries which all
also provide a .cmake file for a cmake find_package(). However the
libInstPatch we are going to install only provides a .pc file and no
.cmake file. Therefore this change makes sure that libraries will be
prepared in /Users/runner/musescore_deps_macos/ and not in
/Users/runner/work/musescore_deps/musescore_deps/musescore_deps_macos/.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>